requirements.txt
aimtorch==0.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
google-auth==2.11.0
google-auth-oauthlib==0.5.2
gspread==5.5.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tweepy==4.10.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
